+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 
+- `ICU4X::PluralRules#select_range` method for plural category selection on number ranges (#106)
 - Allow data gems to be required by gem name (#104)
 
 ## [0.8.1] - 2026-01-12