+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 - `ICU4X::RelativeTimeFormat#format_to_parts` method for breaking down formatted output into typed parts (#117)
 - `ICU4X::ListFormat#format_to_parts` method for breaking down formatted output into typed parts (#116)
-- `ICU4X::DateTimeFormat#format_to_parts` method for breaking down formatted output into typed parts (#114)
 - `ICU4X::NumberFormat#format_to_parts` method for breaking down formatted output into typed parts (#115)
+- `ICU4X::DateTimeFormat#format_to_parts` method for breaking down formatted output into typed parts (#114)
 - `ICU4X::FormattedPart` data class for representing formatted parts (#113)
 - `ICU4X::PluralRules#select_range` method for plural category selection on number ranges (#106)
 - Allow data gems to be required by gem name (#104)